változtatások3

+++ src/main/java/hu/computersSimpleAPI/service/ComputerService.java
@@ -3,7 +3,9 @@ package hu.computersSimpleAPI.service;
 import hu.computersSimpleAPI.domain.Computer;
 import hu.computersSimpleAPI.repository.ComputerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -20,8 +22,9 @@ public class ComputerService {
 
     public Computer getComputerBySerial(int serial) {
         Optional<Computer> computer = repository.findById(serial);
-        System.out.println("Optional: "+computer.get().getMadein());
+        if(computer.isPresent())
         return computer.get();
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
 
     public List<Computer> getComputersByManufacture(String manufacture) {
@@ -32,6 +35,9 @@ public class ComputerService {
     }
 
     public void deleteComputer(int id) {
-       repository.deleteById(id);
+        Optional<Computer> computer = repository.findById(id);
+        if (computer.isPresent())
+            repository.deleteById(id);
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
 }
